restructure addfolder and syncfolder commands
